requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
starlette
slowapi
joblib
python-jose
supabase
numpy
pandas
scikit-learn
python-multipart
gunicorn
requests
